+++ R/ensemble_average.R
@@ -26,7 +26,7 @@
 #' library(tidymodels)
 #' library(modeltime)
 #' library(modeltime.ensemble)
-#' library(tidyverse)
+#' library(dplyr)
 #' library(timetk)
 #'
 #' # Make an ensemble from a Modeltime Table
@@ -89,14 +89,12 @@ print.mdl_time_ensemble_avg <- function(x, ...) {
     print(cli::rule("Modeltime Ensemble", width = min(65, cli::console_width())))
 
     if (x$parameters$type == "mean") {
-        msg <- glue::glue("Ensemble of {x$n_models} Models (MEAN)")
+        msg <- cli::format_inline("Ensemble of {x$n_models} Models (MEAN)")
     } else {
-        msg <- glue::glue("Ensemble of {x$n_models} Models (MEDIAN)")
+        msg <- cli::format_inline("Ensemble of {x$n_models} Models (MEDIAN)")
     }
 
-    print(msg)
-
-    cli::cat_line()
+    cat(msg, "\n\n")
 
     print(x$model_tbl)