+++ tests/testthat.R
@@ -8,11 +8,9 @@ library(modeltime.resample)
 
 # Model dependencies
 library(xgboost)
-library(earth)
 library(glmnet)
 
 # Core Packages
-library(tidyverse)
 library(timetk)
 library(lubridate)