+++ R/ensemble_model_spec.R
@@ -85,6 +85,7 @@
 #' library(modeltime.ensemble)
 #' library(dplyr)
 #' library(timetk)
+#' library(glmnet)
 #'
 #' # Step 1: Make resample predictions for submodels
 #' resamples_tscv <- training(m750_splits) %>%