+++ NEWS.md
@@ -1,3 +1,5 @@
+# modeltime.ensemble (development version)
+
 # modeltime.ensemble 1.0.3
 
 - Resubmit to CRAN  (following timetk archival)